Flyt eksisterende geojson funktionalitet
